UGM 2023: Advanced Training - Compute to Data updates
This change updates the Dockerfile used for the Jupyter notebook
demonstration for the Compute to Data advanced training module.

An assumption about the iRODS service account's Linux user ID: that it
will always be 999. This adds a build argument for the iRODS service
account user ID and group ID so that the build of the image can be run
like this, from the build context:

docker build \
    --build-arg irods_uid=$(id -u irods) \
    --build-arg irods_gid=$(id -g irods) \
    -t testimages/jupyter-digital-filter .