[#46] replacement for "chown -R -u <UID> -g <GID> /opt/conda"
This shortens the compute-to-data docker image build time by providing a more efficient
option (a straight-out move/copy/delete) than chown -R seems to offer for changing
owners in-place for a large directory tree.  To do this, we avail ourselves of tar's
ability to change the owner amid the recursive copying of the directory tree.