[#46] remove uid/gid dependencies in Jupyter docker image
This makes the image generic so that it may be kept at Docker Hub for
download in the compute to data training, instead of having to lead
audience through a time consuming build step.

To accommodate this, we switch to docker exec as the mode for running
the jupyter notebook as well as chown and other UID/GID related pre-
and post-processing steps.  One side effect is we may now specify
a number of commands, not just one, ie a processing pipeline; and
we can specify which user runs each step.